requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Time-Series & Data Processing
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Data Handling
h5py>=3.9.0
